updating legends component to use palette
